change the structure
