+++ .env.example
@@ -1,0 +1,5 @@
+TELEGRAM_BOT_TOKEN=your_telegram_bot_token
+ATS_API_BASE_URL=https://your-ats-api-url
+GEMINI_API_KEY=your_gemini_api_key
+MICROSOFT_APP_ID=your_microsoft_app_id
+MICROSOFT_APP_PASSWORD=your_microsoft_app_password
