+++ app/main.py
@@ -1,0 +1,18 @@
+import os
+import google.generativeai as genai
+import pypdf
+import docx
+import io
+import json
+import httpx
+import asyncio
+from telegram import Update
+
+from dotenv import load_dotenv
+
+
+
+
+
+
+
