+++ requirements.txt
@@ -1,7 +1,11 @@
-python-telegram-bot
+fastapi
+uvicorn[standard]
 python-dotenv
-google-generativeai
+httpx
 pypdf
 python-docx
-httpx
+google-generativeai
 botbuilder-core
+botbuilder-schema
+telegram
+python-telegram-bot==20.3
